requirements.txt
pandas
numpy
ta
matplotlib
scikit-learn
xgboost
tensorflow
keras
joblib
requests
python-dotenv
psycopg2-binary
streamlit
plotly
binance
schedule
textblob
nltk
transformers
torch
openai
tabulate
numba
